Replaced program components with installer
Deleted CASSYS.exe and CASSYS.xlsm which are now packaged in the
installer. In the installer, CASSYS.exe is now named CASSYS Engine.exe
to prevent confusion with the interface. Interface calling of the engine
checks for both the new name (CASSYS Engine.exe) and the old name
(CASSYS.exe) for backward compatibility. Also implemented minor error
checking for if the user closes the Simulation dialog box mid-way,
causing errors with the pivot table.